LayoutTests/editing/execCommand/remove-format-orphaned-list-item.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div id="test" contenteditable><ul><li>hello</li><li>world</li>CyberKit</ul></div>
<script>

Markup.description('This tests RemoveFormant command on an orphaned list item following non-orphaned list items. CyberKit used to remove non-orphaned list items ("hello world"). You should see "hello world CyberKit" below.');
var text = document.getElementById('test').lastChild.lastChild;
var range = document.createRange();
range.setStart(text, 0);
range.setEnd(text, 6);
window.getSelection().removeAllRanges();
window.getSelection().addRange(range);
document.execCommand('RemoveFormat', false, null);
Markup.dump('test');

</script>
</body>
</html>
