LayoutTests/editing/execCommand/remove-format-orphaned-list-item-expected.txt

This tests RemoveFormant command on an orphaned list item following non-orphaned list items. CyberKit used to remove non-orphaned list items ("hello world"). You should see "hello world CyberKit" below.
| <ul>
|   <li>
|     "hello"
|   <li>
|     "world"
|   "<#selection-anchor>CyberKit<#selection-focus>"
